First working compound animation system with a test dance JSON file, script assets moved and renamed

+++ HistoricalDanceAnimator/Assets/Scripts/Data/JSON/JSONChoreography.cs
@@ -4,8 +4,8 @@ public class JSONChoreography
     public string danceName;
     public string danceType;
     public string musicPath;
-    public int musicBPM;
-    public int musicFirstBeatTime;
+    public float musicBPM;
+    public float musicFirstBeatTime;
     public JSONDancerGroup[] groups;
     public JSONDancerPosition[] formation;
     public JSONDanceAction[] choreography;
